requirements.txt
requests[standard]
python-dotenv[standard]
tinydb[standard]